Favorite Collection Cell
ADD - favorite collection view cell, which will represent each favorite or favorite within the user device.
ADD - tap gesture recognizer to the favorite indicator.
ADD - method to setup the cell, which will get the medial specified detailed (show or movie), such as the poster, while also determining if it's a favorite or not.
ADD - tap gesture tap handler, which will toggle the favorite indicator on/off and insert/remove the favorite from the local database.