Media Table View Cell
ADD - media table view cell, which will represent each movie or show within their own section.
ADD - display genres for the expected medias.
ADD - display posters for the expected medias.
ADD - display titles for the expected medias.
ADD - display the score for the expected medias and assign a color depending on the range of the score.