Cast View Controller
ADD - cast view controller, which will assist on displaying the details of a specific cast of a media.
ADD - retrieving the person details from the api.
ADD - retrieving the person images from the api.
ADD - retrieving the person credits/roles from the api.
ADD - tap gesture recognizer to the back button.
ADD - method to process the person requested and retrieved from the api. assist on setting the values to the user interface components.
ADD - method to process the person images. assist on instantiating a image scroller by passing in the image collection.